refac: move `TranscriptType` to `zk/plonk/halo2`
