+++ tachyon/c/zk/plonk/halo2/BUILD.bazel
@@ -137,7 +137,6 @@ tachyon_cc_binary(
     name = "prover_replay",
     srcs = ["prover_replay.cc"],
     deps = [
-        ":transcript_type",
         "//tachyon/base:logging",
         "//tachyon/base/console",
         "//tachyon/base/files:file_path_flag",
@@ -148,15 +147,7 @@ tachyon_cc_binary(
         "//tachyon/c/zk/plonk/keys:bn254_plonk_proving_key_impl",
         "//tachyon/cc/math/finite_fields:prime_field_conversions",
         "//tachyon/zk/plonk/halo2:constants",
-    ],
-)
-
-tachyon_cc_library(
-    name = "transcript_type",
-    hdrs = ["transcript_type.h"],
-    deps = [
-        "//tachyon/base/flag",
-        "@com_google_absl//absl/strings",
+        "//tachyon/zk/plonk/halo2:transcript_type",
     ],
 )
 
@@ -175,7 +166,6 @@ tachyon_cc_binary(
     srcs = ["verifier_replay.cc"],
     deps = [
         ":bn254_shplonk_verifier",
-        ":transcript_type",
         "//tachyon/base:logging",
         "//tachyon/base/console",
         "//tachyon/base/files:file_path_flag",
@@ -183,6 +173,7 @@ tachyon_cc_binary(
         "//tachyon/base/flag:flag_parser",
         "//tachyon/c/math/elliptic_curves/bn/bn254:fr",
         "//tachyon/c/zk/plonk/keys:bn254_plonk_proving_key_impl",
+        "//tachyon/zk/plonk/halo2:transcript_type",
     ],
 )
