+++ tachyon/zk/plonk/halo2/BUILD.bazel
@@ -222,6 +222,15 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
+    name = "transcript_type",
+    hdrs = ["transcript_type.h"],
+    deps = [
+        "//tachyon/base/flag",
+        "@com_google_absl//absl/strings",
+    ],
+)
+
+tachyon_cc_library(
     name = "verifier",
     hdrs = ["verifier.h"],
     deps = [
